Refactor configuration and type imports for improved clarity; remove obsolete type definitions and update related references
