+++ pr-review/src/io/ai/aiReview.ts
@@ -2,9 +2,8 @@ import * as core from "@actions/core"
 import { ChatMessage } from "@sap-ai-sdk/orchestration"
 import { inspect } from "node:util"
 import parseDiff from "parse-diff"
-import { Config } from "../../domain/model/config.ts"
 import { AiReviewSchema } from "../../domain/model/review.ts"
-import { AiReview, ReviewComment } from "../../domain/model/types.ts"
+import { AiReview, Config, ReviewComment } from "../../domain/model/types.ts"
 import { resolveHunkReferencesInComments } from "../../domain/services/hunk-reader.ts"
 import * as aiCoreClient from "./ai-core-client.ts"
